Patch a snake_bite.
Made sure that timestampable SQL table columns were snake_cased, to the cost of pitfall work if we want to switch to an ODM later.
Which we probably won't anyways.

Procrastination, you say ?

I really dislike having to migrate databases after release, and this just looked BAD in the logs. Now there are no broken windows.